requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
numpy==1.23.4
pandas==1.5.1
py-markdown-table==0.3.3
pyTelegramBotAPI==4.7.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
tabulate==0.9.0
urllib3==1.26.12
